requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langchainhub
faiss-cpu
pymupdf
streamlit
python-dotenv
google-ai-generativelanguage
fastapi
uvicorn
python-multipart

